Encapsulate common logic of ScoreProcessor
